# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface BillingDeviceInput {
  /**
     * @minLength 20
     * @maxLength 100
     */
  deviceId: string;
}

export interface BillingStatusInput {
  /**
     * @minLength 20
     * @maxLength 100
     */
  deviceId: string;
  onboardingComplete: boolean;
}

export type BillingCheckoutInputPlan = typeof BillingCheckoutInputPlan[keyof typeof BillingCheckoutInputPlan];


export const BillingCheckoutInputPlan = {
  monthly: 'monthly',
  annual: 'annual',
} as const;

export interface BillingCheckoutInput {
  /**
     * @minLength 20
     * @maxLength 100
     */
  deviceId: string;
  plan: BillingCheckoutInputPlan;
}

export type BillingStatusAccess = typeof BillingStatusAccess[keyof typeof BillingStatusAccess];


export const BillingStatusAccess = {
  free: 'free',
  trial: 'trial',
  pro: 'pro',
} as const;

export interface BillingStatus {
  access: BillingStatusAccess;
  isPro: boolean;
  trialDaysRemaining: number;
  /** @nullable */
  trialEndsAt: string | null;
  monthlyPrice: string;
  annualPrice: string;
  annualSavingPercent: number;
  canManage: boolean;
}

export interface BillingRedirect {
  url: string;
}