# lib/api-zod/src/generated/types/billingCheckoutInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BillingCheckoutInputPlan } from './billingCheckoutInputPlan';

export interface BillingCheckoutInput {
  /**
     * @minLength 20
     * @maxLength 100
     */
  deviceId: string;
  plan: BillingCheckoutInputPlan;
}