# lib/api-zod/src/generated/types/billingDeviceInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface BillingDeviceInput {
  /**
     * @minLength 20
     * @maxLength 100
     */
  deviceId: string;
}