# lib/api-zod/src/generated/types/billingStatus.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BillingStatusAccess } from './billingStatusAccess';

export interface BillingStatus {
  access: BillingStatusAccess;
  isPro: boolean;
  trialDaysRemaining: number;
  /** @nullable */
  trialEndsAt: Date | null;
  monthlyPrice: string;
  annualPrice: string;
  annualSavingPercent: number;
  canManage: boolean;
}