# lib/api-zod/src/generated/types/billingStatusInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface BillingStatusInput {
  /**
     * @minLength 20
     * @maxLength 100
     */
  deviceId: string;
  onboardingComplete: boolean;
}